pyproject.toml
[tool.poetry]
name = "netdiscoverit"
version = "0.1.0"
description = "Network discovery and ITIL documentation tool"
authors = ["Chris Cline <Chris.Cline@openether.net>"]
license = "MIT"
readme = "README.md"
packages = [{include = "netdiscoverit", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.9.0"  # CLI framework
python-nmap = "^0.7.1"  # Network scanning
scapy = "^2.5.0"  # Packet crafting
pysnmp = "^5.0.0"  # Device identification
pandas = "^2.0.0"  # Data categorization
jinja2 = "^3.1.0"  # Templating for docs
reportlab = "^4.0.0"  # PDF generation
networkx = "^3.1"  # Graph analysis
pydot = "^1.4.0"  # Graphviz integration
pyyaml = "^6.0"  # Config/outputs
pydantic = "^2.4"  # Data validation
pluggy = "^1.3"  # Plugin system
asyncio = "*"  # Built-in, for async

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
flake8 = "^6.0.0"
bandit = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
netdiscoverit = "netdiscoverit.main:cli"